Remove Redundant tet cases
